Refactored Test
