+++ src/main/java/com/company/functional/Main.java
@@ -1,0 +1,9 @@
+package com.company.functional;
+
+import java.util.Map;
+
+public class Main {
+
+    public static Map<String,Integer> getWordsWithOccurrenceInAscendingOrder(String inputString) {
+    }
+}
